pyproject.toml
[tool.poetry]
name = "discord"
version = "0.1.0"
description = "The discord client for MiceBot."
authors = ["MiceBot Collaborators."]
license = "MIT"
repository = "https://github.com/micebot/server.git"

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = "^1.3.3"
httpx = "^0.13.3"
pydantic = "^1.6.1"

[tool.poetry.dev-dependencies]
pylama = "^7.7.1"
coverage = "^5.1"
ipdb = "^0.13.3"
faker = "^4.1.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
